pyproject.toml
[project]
name = "winter-camp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.9.6",
    "chromadb>=1.4.1",
    "docx>=0.2.4",
    "fastapi>=0.128.2",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.7",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "tqdm>=4.67.3",
    "uvicorn>=0.40.0",
]
